# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost","description":"Generated server url"}],"tags":[{"name":"Admin Common Controller","description":"模型管理"},{"name":"Auth Controller","description":"用户注册登录接口"},{"name":"Admin Model Controller","description":"模型管理"},{"name":"Vendor Controller","description":"服务商相关接口"},{"name":"User Controller","description":"用户信息相关接口"},{"name":"Admin Order Controller","description":"管理员订单管理"},{"name":"Api Key Controller","description":"用户API Key的增删改查接口"}],"paths":{"/be-admin/vendors/{vendorId}/tokens/{tokenId}":{"put":{"tags":["Vendor Controller"],"operationId":"update","parameters":[{"name":"vendorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"tokenId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VendorTokenUpdateReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoBoolean"}}}}}}},"/be-admin/models/{id}":{"put":{"tags":["Admin Model Controller"],"summary":"Update Model","description":"更新模型信息","operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelUpdateReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoString"}}}}}}},"/be/keys":{"get":{"tags":["Api Key Controller"],"summary":"Get Api Keys","description":"获取用户Api Key列表","operationId":"list","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoListApiKeyResp"}}}}}},"post":{"tags":["Api Key Controller"],"summary":"Create Api Key","description":"为用户创建Api Key","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiKeyCreateReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoVoid"}}}}}}},"/be/auth/register":{"post":{"tags":["Auth Controller"],"summary":"Register","description":"用户注册","operationId":"register","requestBody":{"description":"注册表单数据","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterReq"}}},"required":true},"responses":{"200":{"description":"注册成功，返回JWT Token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"},"examples":{"success":{"summary":"注册成功","description":"success","value":{"code":200,"desc":"注册成功","result":"eyJhbGciOiJIUzI1NiJ9.eyJ1c2...."}}}}}}}}},"/be/auth/login":{"post":{"tags":["Auth Controller"],"summary":"Login","description":"用户登录","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoString"}}}}}}},"/be-admin/vendors":{"get":{"tags":["Vendor Controller"],"summary":"Get Vendor List","description":"服务商列表查询","operationId":"vendors","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoListVendorResp"}}}}}},"post":{"tags":["Vendor Controller"],"summary":"Create New Vendor","description":"创建新服务商","operationId":"createVendor","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VendorCreateReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoString"}}}}}}},"/be-admin/vendors/{vendorId}/tokens":{"post":{"tags":["Vendor Controller"],"operationId":"createTokens","parameters":[{"name":"vendorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VendorTokenCreateReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoBoolean"}}}}}}},"/be-admin/models":{"get":{"tags":["Admin Model Controller"],"summary":"Get Models","description":"获取模型详细信息，相比用户端接口，多了模型映射关系","operationId":"list_1","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoListModelResp"}}}}}},"post":{"tags":["Admin Model Controller"],"summary":"Create New Model","description":"创建新模型，并选定服务商模型与该模型绑定。","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelCreateReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoString"}}}}}}},"/api/v1/chat/completions":{"post":{"tags":["api-controller"],"operationId":"syncChat","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommonChatReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}},"text/event-stream":{"schema":{"type":"object"}}}}}}},"/be-admin/models/{id}/status":{"patch":{"tags":["Admin Model Controller"],"summary":"Toggle Model Status","description":"启用或禁用模型。禁用后，用户将无法调用该模型。","operationId":"changeStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelStatusUpdateReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoBoolean"}}}}}}},"/be/user/profile":{"get":{"tags":["User Controller"],"operationId":"profile","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoUserProfileResp"}}}}}}},"/be/plans":{"get":{"tags":["plan-controller"],"operationId":"list_2","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoListRechargePlanResp"}}}}}}},"/be/models":{"get":{"tags":["model-controller"],"operationId":"models","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoListModelResp"}}}}}}},"/be/models/types":{"get":{"tags":["model-controller"],"operationId":"modelTypes","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoListMapStringString"}}}}}}},"/be/billing":{"get":{"tags":["billing-controller"],"operationId":"list_3","parameters":[{"name":"req","in":"query","required":true,"schema":{"$ref":"#/components/schemas/BillQueryReq"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoIPageBillVO"}}}}}}},"/be/billing/stats":{"get":{"tags":["billing-controller"],"operationId":"stats","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoBillStatsDto"}}}}}}},"/be-admin/vendors/types":{"get":{"tags":["Vendor Controller"],"operationId":"vendorTypes","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoListMapStringString"}}}}}}},"/be-admin/vendors/models":{"get":{"tags":["Vendor Controller"],"summary":"Get All Vendor Models","description":"获取全部销售商的全部模型","operationId":"models_1","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoListVendorModelResp"}}}}}}},"/be-admin/orders":{"get":{"tags":["Admin Order Controller"],"summary":"Get Orders","description":"获取全部用户的API订单记录","operationId":"list_4","parameters":[{"name":"pageNum","in":"query","description":"当前页码","required":false,"schema":{"type":"integer","format":"int32","default":1,"description":"当前页码","example":1},"example":1},{"name":"pageSize","in":"query","description":"每页条数","required":false,"schema":{"type":"integer","format":"int32","default":10,"description":"每页条数","example":10},"example":10},{"name":"startDate","in":"query","description":"开始时间","required":false,"schema":{"type":"string","format":"date-time","description":"开始时间","example":"2026-01-17 00:00:00"},"example":"2026-01-17 00:00:00"},{"name":"endDate","in":"query","description":"结束时间","required":false,"schema":{"type":"string","format":"date-time","description":"结束时间","example":"2026-01-31 23:59:59"},"example":"2026-01-31 23:59:59"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoIPageApiOrderResp"}}}}}}},"/be-admin/common/dashboard":{"get":{"tags":["Admin Common Controller"],"operationId":"dashboardQuery","parameters":[{"name":"date","in":"query","required":false,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoDashboardResp"}}}}}}},"/be-admin/common/check":{"get":{"tags":["Admin Common Controller"],"summary":"Check auth","description":"检查用户是否具有管理员权限","operationId":"check","responses":{"200":{"description":"用户具有admin权限，返回ok","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"},"examples":{"success":{"summary":"成功","description":"success","value":{"code":200,"desc":"Success","result":"ok"}}}}}},"403":{"description":"权限不足，拒绝访问","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"},"examples":{"forbidden":{"summary":"权限不足","description":"forbidden","value":{"code":403,"result":"权限不足"}}}}}}}}},"/be/keys/{id}":{"delete":{"tags":["Api Key Controller"],"summary":"Delete Api Key","description":"删除选择的Api Key","operationId":"delete","parameters":[{"name":"id","in":"path","description":"Api Key ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDtoVoid"}}}}}}}},"components":{"schemas":{"VendorTokenUpdateReq":{"type":"object","properties":{"apiKey":{"type":"string"},"tokenName":{"type":"string"},"isActive":{"type":"integer","format":"int64"},"expiresAt":{"type":"string","format":"date-time"}}},"ResponseDtoBoolean":{"type":"object","description":"通用返回体","properties":{"code":{"type":"integer","format":"int32","description":"状态码","example":200},"desc":{"type":"string","description":"状态描述","example":"success"},"result":{"type":"boolean","description":"结果"}},"required":["code"]},"ModelUpdateReq":{"type":"object","description":"更新模型请求对象","properties":{"name":{"type":"string","description":"模型名称(唯一标识)","example":"deepseek-chat","minLength":1},"displayName":{"type":"string","description":"模型展示名称","example":"DeepSeek-V3.2","minLength":1},"isFree":{"type":"boolean","description":"是否免费","example":false},"category":{"type":"string","description":"模型种类","enum":["chat","image","audio","video","embedding"],"example":"chat"},"platformType":{"type":"string","description":"模型厂商","enum":["DEEP_SEEK"],"example":"DEEP_SEEK"},"description":{"type":"string","description":"模型描述","example":"DeepSeek推出的首个将思考融入工具使用的模型，并且同时支持思考模式与非思考模式的工具调用，智能水平与 GPT-5 相当"},"vendorModelIds":{"type":"array","description":"绑定的服务商模型(Vendor Model)ID列表","items":{"type":"integer","format":"int64"},"minItems":1},"pricingItems":{"type":"array","description":"计费项配置列表 (当 isFree 为 false 时必填)","items":{"$ref":"#/components/schemas/PricingItemCreateReq"}}},"required":["category","displayName","isFree","name","platformType","vendorModelIds"]},"PricingItemCreateReq":{"type":"object","description":"计费项","properties":{"unit":{"type":"string","description":"计费单位","enum":["mtokens","mchars","times","nums","minutes","seconds","none"],"example":"mtokens"},"priceInput":{"type":"number","description":"输入单价","example":1.5,"minimum":0.0},"priceOutput":{"type":"number","description":"输出单价","example":3.0,"minimum":0.0}},"required":["priceInput","priceOutput","unit"]},"ResponseDtoString":{"type":"object","description":"通用返回体","properties":{"code":{"type":"integer","format":"int32","description":"状态码","example":200},"desc":{"type":"string","description":"状态描述","example":"success"},"result":{"type":"string","description":"结果"}},"required":["code"]},"ApiKeyCreateReq":{"type":"object","description":"API Key创建请求","properties":{"keyName":{"type":"string","description":"API Key名称","example":"test api key"},"quota":{"type":"number","description":"限额","example":100.0},"expireTime":{"type":"string","format":"date-time","description":"过期时间","example":"2026-01-08T00:00:00"}},"required":["expireTime","keyName","quota"]},"ResponseDtoVoid":{"type":"object","description":"通用返回体","properties":{"code":{"type":"integer","format":"int32","description":"状态码","example":200},"desc":{"type":"string","description":"状态描述","example":"success"},"result":{"description":"结果"}},"required":["code"]},"RegisterReq":{"type":"object","description":"用户注册请求","properties":{"username":{"type":"string","description":"用户名","example":"example_username","maxLength":20,"minLength":4},"password":{"type":"string","description":"密码","example":"example_password","maxLength":30,"minLength":8},"email":{"type":"string","format":"email","description":"邮箱","example":"example@panapi.com"}},"required":["email","password","username"]},"ResponseDto":{"type":"object","description":"通用返回体","properties":{"code":{"type":"integer","format":"int32","description":"状态码","example":200},"desc":{"type":"string","description":"状态描述","example":"success"},"result":{"description":"结果"}},"required":["code"]},"LoginReq":{"type":"object","properties":{"identifier":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["identifier","password"]},"VendorCreateReq":{"type":"object","description":"服务商创建请求","properties":{"name":{"type":"string","description":"服务商名称","example":"DeepSeek-高可用渠道"},"apiBaseUrl":{"type":"string","description":"API基础地址","example":"https://api.deepseek.com/"},"venType":{"type":"string","description":"服务商类型","enum":["COMMON","DEEP_SEEK","GLM","DOU_BAO","XUN_FEI"],"example":"DEEP_SEEK"}},"required":["apiBaseUrl","name","venType"]},"VendorTokenCreateReq":{"type":"object","properties":{"apiKey":{"type":"string","minLength":1},"tokenName":{"type":"string","minLength":1},"isActive":{"type":"integer","format":"int64"},"expiresAt":{"type":"string","format":"date-time"}},"required":["apiKey","tokenName"]},"ModelCreateReq":{"type":"object","description":"创建模型请求对象","properties":{"name":{"type":"string","description":"模型名称(唯一标识)","example":"deepseek-chat","minLength":1},"displayName":{"type":"string","description":"模型展示名称","example":"DeepSeek-V3.2","minLength":1},"isFree":{"type":"boolean","description":"是否免费","example":false},"category":{"type":"string","description":"模型种类","enum":["chat","image","audio","video","embedding"],"example":"chat"},"platformType":{"type":"string","description":"模型厂商","enum":["DEEP_SEEK"],"example":"DEEP_SEEK"},"description":{"type":"string","description":"模型描述","example":"DeepSeek推出的首个将思考融入工具使用的模型，并且同时支持思考模式与非思考模式的工具调用，智能水平与 GPT-5 相当"},"vendorModelIds":{"type":"array","description":"绑定的服务商模型(Vendor Model)ID列表","items":{"type":"integer","format":"int64"},"minItems":1},"pricingItems":{"type":"array","description":"计费项配置列表 (当 isFree 为 false 时必填)","items":{"$ref":"#/components/schemas/PricingItemCreateReq"}}},"required":["category","displayName","isFree","name","platformType","vendorModelIds"]},"CommonChatReq":{"type":"object","properties":{"messages":{"type":"array","items":{"$ref":"#/components/schemas/Message"}},"model":{"type":"string"},"logprobs":{"type":"boolean"},"metadata":{"type":"object","additionalProperties":{"type":"string"}},"modalities":{"type":"array","items":{"type":"string"}},"n":{"type":"integer","format":"int32"},"prediction":{},"stop":{},"stream":{"type":"boolean"},"temperature":{"type":"number","format":"double"},"tools":{"type":"array","items":{"$ref":"#/components/schemas/Tool"}},"user":{"type":"string"},"seed":{"type":"integer","format":"int32"},"frequency_penalty":{"type":"number","format":"double"},"logit_bias":{"type":"object","additionalProperties":{"type":"integer","format":"int32"}},"max_completion_tokens":{"type":"integer","format":"int32"},"max_tokens":{"type":"integer","format":"int32"},"parallel_tool_calls":{"type":"boolean"},"presence_penalty":{"type":"number","format":"double"},"prompt_cache_key":{"type":"string"},"reasoning_effort":{"type":"string"},"response_format":{"$ref":"#/components/schemas/ResponseFormat"},"service_tier":{"type":"string"},"store":{"type":"boolean"},"stream_options":{"$ref":"#/components/schemas/StreamOptions"},"tool_choice":{},"top_logprobs":{"type":"integer","format":"int32"},"top_p":{"type":"number","format":"double"},"verbosity":{"type":"string"}},"required":["messages","model"]},"ContentPart":{"type":"object","properties":{"type":{"type":"string"},"text":{"type":"string"},"image_url":{"$ref":"#/components/schemas/ImageUrl"}}},"CustomTool":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"format":{}}},"FunctionCall":{"type":"object","properties":{"name":{"type":"string"},"arguments":{"type":"string"}}},"FunctionTool":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"parameters":{},"strict":{"type":"boolean"}}},"ImageUrl":{"type":"object","properties":{"url":{"type":"string"}}},"JsonSchema":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"schema":{},"strict":{"type":"boolean"}}},"Message":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/ContentPart"}},"role":{"type":"string"},"name":{"type":"string"},"tool_calls":{"type":"array","items":{"$ref":"#/components/schemas/ToolCall"}},"tool_call_id":{"type":"string"}},"required":["content","role"]},"ResponseFormat":{"type":"object","properties":{"type":{"type":"string"},"json_schema":{"$ref":"#/components/schemas/JsonSchema"}}},"StreamOptions":{"type":"object","properties":{"include_obfuscation":{"type":"boolean"},"include_usage":{"type":"boolean"}}},"Tool":{"type":"object","properties":{"type":{"type":"string"},"function":{"$ref":"#/components/schemas/FunctionTool"},"custom":{"$ref":"#/components/schemas/CustomTool"}}},"ToolCall":{"type":"object","properties":{"id":{"type":"string"},"type":{"type":"string"},"function":{"$ref":"#/components/schemas/FunctionCall"}}},"ModelStatusUpdateReq":{"type":"object","properties":{"enabled":{"type":"boolean","description":"状态","example":true}},"required":["enabled"]},"ApiKeyResp":{"type":"object","description":"Api Key 响应对象","properties":{"id":{"type":"integer","format":"int64","description":"Api Key ID","example":11},"keyName":{"type":"string","description":"Api Key 名称","example":"TEST KEY"},"apiKey":{"type":"string","description":"Api Key Token","example":"07471tyc849u23382d9b07539741cfe4"},"quota":{"type":"number","description":"限额额度","example":5000.0},"createTime":{"type":"string","format":"date-time","description":"创建时间","example":"2025-11-24 18:32:54"}}},"ResponseDtoUserProfileResp":{"type":"object","description":"通用返回体","properties":{"code":{"type":"integer","format":"int32","description":"状态码","example":200},"desc":{"type":"string","description":"状态描述","example":"success"},"result":{"$ref":"#/components/schemas/UserProfileResp","description":"结果"}},"required":["code"]},"UserProfileResp":{"type":"object","properties":{"apiKeys":{"type":"array","items":{"$ref":"#/components/schemas/ApiKeyResp"}},"user":{"$ref":"#/components/schemas/UserResp"},"todayTokens":{"type":"integer","format":"int64"},"todayRequests":{"type":"integer","format":"int64"}}},"UserResp":{"type":"object","properties":{"username":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"balance":{"type":"number"},"createTime":{"type":"string","format":"date-time"}}},"RechargePlanResp":{"type":"object","properties":{"planId":{"type":"integer","format":"int64"},"name":{"type":"string"},"price":{"type":"number"},"desc":{"type":"string"},"displayMd":{"type":"string"},"isRecommend":{"type":"integer","format":"int32"},"sortOrder":{"type":"integer","format":"int32"},"status":{"type":"integer","format":"int32"}}},"ResponseDtoListRechargePlanResp":{"type":"object","description":"通用返回体","properties":{"code":{"type":"integer","format":"int32","description":"状态码","example":200},"desc":{"type":"string","description":"状态描述","example":"success"},"result":{"type":"array","description":"结果","items":{"$ref":"#/components/schemas/RechargePlanResp"}}},"required":["code"]},"BindingDetail":{"type":"object","description":"模型绑定详情","properties":{"id":{"type":"integer","format":"int64","description":"绑定关系ID","example":1001},"modelId":{"type":"integer","format":"int64","description":"平台模型ID","example":1},"venModelId":{"type":"integer","format":"int64","description":"服务商原始模型ID（关联 ven_model 表）","example":50},"enabled":{"type":"integer","format":"int32","description":"绑定是否启用：0-禁用, 1-启用","example":1}}},"ModelResp":{"type":"object","description":"模型展示详情对象（包含映射关系）","properties":{"id":{"type":"integer","format":"int64","description":"模型ID","example":1},"name":{"type":"string","description":"模型唯一标识名称","example":"gpt-4-turbo"},"displayName":{"type":"string","description":"对外展示的模型名称","example":"ChatGPT 4 Turbo"},"isFree":{"type":"boolean","description":"是否免费","example":false},"category":{"type":"string","description":"模型类别：chat-对话, image-绘画, audio-语音, video-视频, embedding-向量","enum":["chat","image","audio","video","embedding"]},"description":{"type":"string","description":"模型描述信息","example":"这是目前最强大的语言模型之一"},"isActive":{"type":"integer","format":"int32","description":"状态：0-禁用, 1-启用","example":1},"bindings":{"type":"array","description":"绑定的服务商模型映射列表(用户端不包含)","items":{"$ref":"#/components/schemas/BindingDetail"}},"pricingItems":{"type":"array","description":"模型计费项列表","items":{"$ref":"#/components/schemas/PricingItemResp"}}}},"PricingItemResp":{"type":"object","description":"模型计费项详情","properties":{"id":{"type":"integer","format":"int64","description":"计费项ID","example":100},"unit":{"type":"string","description":"计费单位","enum":["mtokens","mchars","times","nums","minutes","seconds","none"],"example":"mtokens"},"priceInput":{"type":"number","description":"输入单价","example":0.01},"priceOutput":{"type":"number","description":"输出单价","example":0.02},"currency":{"type":"string","description":"货币类型","example":"CNY"},"isActive":{"type":"integer","format":"int32","description":"状态：0-禁用, 1-启用","example":1}}},"ResponseDtoListModelResp":{"type":"object","description":"通用返回体","properties":{"code":{"type":"integer","format":"int32","description":"状态码","example":200},"desc":{"type":"string","description":"状态描述","example":"success"},"result":{"type":"array","description":"结果","items":{"$ref":"#/components/schemas/ModelResp"}}},"required":["code"]},"ResponseDtoListMapStringString":{"type":"object","description":"通用返回体","properties":{"code":{"type":"integer","format":"int32","description":"状态码","example":200},"desc":{"type":"string","description":"状态描述","example":"success"},"result":{"type":"array","description":"结果","items":{"type":"object","additionalProperties":{"type":"string"}}}},"required":["code"]},"ResponseDtoListApiKeyResp":{"type":"object","description":"通用返回体","properties":{"code":{"type":"integer","format":"int32","description":"状态码","example":200},"desc":{"type":"string","description":"状态描述","example":"success"},"result":{"type":"array","description":"结果","items":{"$ref":"#/components/schemas/ApiKeyResp"}}},"required":["code"]},"BillQueryReq":{"type":"object","properties":{"billType":{"type":"string"},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"},"pageNum":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"}}},"BillVO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"billType":{"type":"string"},"amount":{"type":"number"},"relatedId":{"type":"string"},"description":{"type":"string"},"modelTagId":{"type":"integer","format":"int64"},"modelTagName":{"type":"string"},"createTime":{"type":"string","format":"date-time"}}},"IPageBillVO":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"current":{"type":"integer","format":"int64"},"records":{"type":"array","items":{"$ref":"#/components/schemas/BillVO"}},"pages":{"type":"integer","format":"int64","deprecated":true},"total":{"type":"integer","format":"int64"}}},"ResponseDtoIPageBillVO":{"type":"object","description":"通用返回体","properties":{"code":{"type":"integer","format":"int32","description":"状态码","example":200},"desc":{"type":"string","description":"状态描述","example":"success"},"result":{"$ref":"#/components/schemas/IPageBillVO","description":"结果"}},"required":["code"]},"BillStats":{"type":"object","properties":{"totalAmount":{"type":"number"},"transactionCount":{"type":"integer","format":"int64"}}},"BillStatsDto":{"type":"object","properties":{"today":{"$ref":"#/components/schemas/BillStats"},"month":{"$ref":"#/components/schemas/BillStats"}}},"ResponseDtoBillStatsDto":{"type":"object","description":"通用返回体","properties":{"code":{"type":"integer","format":"int32","description":"状态码","example":200},"desc":{"type":"string","description":"状态描述","example":"success"},"result":{"$ref":"#/components/schemas/BillStatsDto","description":"结果"}},"required":["code"]},"ResponseDtoListVendorResp":{"type":"object","description":"通用返回体","properties":{"code":{"type":"integer","format":"int32","description":"状态码","example":200},"desc":{"type":"string","description":"状态描述","example":"success"},"result":{"type":"array","description":"结果","items":{"$ref":"#/components/schemas/VendorResp"}}},"required":["code"]},"VendorResp":{"type":"object","description":"服务商 响应对象","properties":{"id":{"type":"integer","format":"int64","description":"服务商ID","example":5},"name":{"type":"string","description":"服务商名称","example":"OpenAI"},"apiBaseUrl":{"type":"string","description":"API基础地址","example":"https://api.openai.com/v1"},"venType":{"type":"string","description":"服务商类型","enum":["COMMON","DEEP_SEEK","GLM","DOU_BAO","XUN_FEI"],"example":"OPENAI"},"tokens":{"type":"array","items":{"$ref":"#/components/schemas/VendorTokenReq"}}}},"VendorTokenReq":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"vendorId":{"type":"integer","format":"int64"},"apiKey":{"type":"string"},"tokenName":{"type":"string"},"isActive":{"type":"boolean"},"expiresAt":{"type":"string","format":"date-time"}}},"ResponseDtoListVendorModelResp":{"type":"object","description":"通用返回体","properties":{"code":{"type":"integer","format":"int32","description":"状态码","example":200},"desc":{"type":"string","description":"状态描述","example":"success"},"result":{"type":"array","description":"结果","items":{"$ref":"#/components/schemas/VendorModelResp"}}},"required":["code"]},"VendorModelResp":{"type":"object","description":"服务商模型响应对象","properties":{"id":{"type":"integer","format":"int64","description":"服务商模型ID (ven_model 表 ID)","example":14},"vendorId":{"type":"integer","format":"int64","description":"服务商ID (关联 vendor 表)","example":2},"name":{"type":"string","description":"服务商内部模型名称 (如: gpt-4, deepseek-chat)","example":"gpt-4-0125-preview"}}},"ApiOrderResp":{"type":"object","description":"API 订单响应对象","properties":{"reqId":{"type":"string","description":"请求id","example":2008073209202606080},"totalTokens":{"type":"integer","format":"int32","description":"总 Token 数","example":1024},"callTime":{"type":"string","format":"date-time","description":"API 调用发起时间","example":"2026-01-05 15:09:30"},"userId":{"type":"integer","format":"int64","description":"用户 ID","example":18},"amount":{"type":"number","description":"用户实际扣费金额","example":0.1575},"cost":{"type":"number","description":"平台向服务商支付的成本金额","example":0.115},"modelName":{"type":"string","description":"调用的平台模型标识名称","example":"gpt-image-1"},"apiType":{"type":"string","description":"API 请求路径/接口类型","example":"/v1/chat/completions"}}},"IPageApiOrderResp":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"current":{"type":"integer","format":"int64"},"records":{"type":"array","items":{"$ref":"#/components/schemas/ApiOrderResp"}},"pages":{"type":"integer","format":"int64","deprecated":true},"total":{"type":"integer","format":"int64"}}},"ResponseDtoIPageApiOrderResp":{"type":"object","description":"通用返回体","properties":{"code":{"type":"integer","format":"int32","description":"状态码","example":200},"desc":{"type":"string","description":"状态描述","example":"success"},"result":{"$ref":"#/components/schemas/IPageApiOrderResp","description":"结果"}},"required":["code"]},"DashboardResp":{"type":"object","properties":{"userCount":{"type":"integer","format":"int32"},"orderCount":{"type":"integer","format":"int32"},"revenue":{"type":"number"}}},"ResponseDtoDashboardResp":{"type":"object","description":"通用返回体","properties":{"code":{"type":"integer","format":"int32","description":"状态码","example":200},"desc":{"type":"string","description":"状态描述","example":"success"},"result":{"$ref":"#/components/schemas/DashboardResp","description":"结果"}},"required":["code"]}},"securitySchemes":{"BearerAuth":{"type":"http","description":"基于Bearer Token形式的鉴权，用于/api下所有接口","scheme":"bearer","bearerFormat":"JWT"},"CookieJwtAuth":{"type":"apiKey","description":"用于前端接口的JWT形式鉴权 /be /be-admin","name":"usif","in":"cookie"}}}}